workers: Don't let core spawn workers
There are two sensible mechanisms for getting workers running:

 a) Let the core manage workers by forking out new worker processes, and
 manage the pool.
 b) Let workers spawn themselves and register with the master process.

Naemon looked at the two sensible approaches and chose "both and
neither" - the core spawns initial workers, but it cannot respawn them;
the workers register with the master, not the other way around.

Killing a bunch of code gets us to b, and now it's your init system's
problem to actually launch/manage workers.

Signed-off-by: Robin Sonefors <ozamosi@flukkost.nu>