Get rid of check delay/spread
Everything about it is just absolutely, ridonculously daft - just run
the bloody checks normally, and stop dicking about with whatever that
was!

The most hilarious thing about it is that it *does* have the correct
code - too - except only as a fallback, in case the insane code ended up
with a crazy value.

I also kicked out the whole scheduling_info struct, because it is all
just silly - I know sysadmins like to see numbers, and I realize there's
a psychological advantage to seeing that my "SMART" "Service interleave
factor" is bigger than yours, but your mental health is not my problem,
I am not a psychiatrist, and this is not medical advice.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>