defaults: Default syslogging to off
Why? Well, emscripten doesn't implement the syslog() function, and it's
called before the configuration has been parsed to determine whether to
actually syslog or not, so I can't provide a config that defaults to
off.

This also aligns the default value when parsing the config, to the
default value in the sample config. Nifty, huh?

Signed-off-by: Robin Sonefors <ozamosi@flukkost.nu>