iobroker: Rewrite poll backend to work
It's still horrifically inefficient, space-wise, but at least it's not
quite as inefficient, time-wise, and at least it actually does what it
says on the label.

The select backend is still utterly, 100% fubar, and I'm not even gonna
bother. Looking at it, the most merciful thing to do seems to be to take
it out back and shoot it.

Signed-off-by: Robin Sonefors <ozamosi@flukkost.nu>