nsock: rewrite to be fully asynchronous
That includes the connect() call, which was previously always
synchronous.

Making connect asynchronous as well would mean reconnects can be done
without extra effort.

In order to be that, we need to return something inbetween we create the
socket, and we actually connect, so the connector can add an iobroker
listener before the conectee has a chance to accept the connector.

But That's Not All: if you order now, you'll get TCP sockets (and hence
TCP workers, query handlers) for free!

Signed-off-by: Robin Sonefors <ozamosi@flukkost.nu>