hash: Get rid of the remaining hash implementation
Oh, look, there already is a generic library that implements common
algorithms in reusable ways! I know, let's not reinvent!

Using glib has the nice side-effect of the hashtable actually being,
like, useful and feature-rich, and has a sane view on memory management.
And not that this is a big new dep - your bloody init system probably
even depends on it.

Not that naemon really works with a sane memory model - my god, the
query handler API is incompatible with any memory that Isn't leaked‽ -
but still, it could be used for greatness in the future.

Signed-off-by: Robin Sonefors <ozamosi@flukkost.nu>